pyproject.toml
[project]
name = "veterans-community-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "black>=26.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "minio>=7.2.20",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sqladmin>=0.23.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["veterans_api*"]
exclude = ["tests*", "test*"]  # Добавьте исключение тестов

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml -v"
asyncio_mode = "auto"  # Важно для async тестов
filterwarnings = [
    "error",  # Превращать warnings в errors
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
